feat(website): give the docs the landing's language and its nav
The two halves of the site did not look related. The landing is pure black with
square hairline boxes, uppercase micro-labels and a rule under every section.
The docs were a shade off black with rounded cards, softer borders and their
own header, so crossing between them felt like leaving.

The nav is now one component both render, the landing's own, unchanged, in the
same 1320px column on both. The markup is byte-identical either side, so the
wordmark, the buttons and the rule under them do not move when you cross. That
is what makes it read as one site rather than two: react.doctor does the same
thing, one persistent header, and no route animation at all.

The rest is the landing's vocabulary applied to the docs. Every rounded corner
in the docs chrome is gone, eleven of them. Containers take the figure border,
white/30, structural rules take white/15, and every h2 sits on one. The ground
is pure black, which globals.css previously argued against in a comment that is
now also gone. Sidebar groups and the wordmark use the landing's label, bold
11px uppercase on 0.08em, and the sidebar items drop to its 13px body size.

pr-review keeps its rounded pills. They imitate GitHub's own UI rather than
following ours.
