+++ .changeset/docs-landing-style.md
@@ -1,0 +1,5 @@
+---
+---
+
+Website only. The docs adopt the landing's design language and share its nav.
+No published package changes.
