+++ packages/website/src/app/docs/layout.tsx
@@ -1,6 +1,6 @@
-import { DocsHeader } from "@/components/docs/docs-header";
 import { Sidebar } from "@/components/docs/sidebar";
 import { Toc } from "@/components/docs/toc";
+import { Nav } from "@/components/site-nav";
 
 export default function DocsLayout({
 	children,
@@ -8,9 +8,11 @@ export default function DocsLayout({
 	children: React.ReactNode;
 }) {
 	return (
-		<div className="flex min-h-screen flex-col bg-[#0a0a0a] font-mono text-neutral-300">
-			<DocsHeader />
-			<div className="flex flex-1">
+		<div className="flex min-h-screen flex-col bg-black font-mono text-neutral-300">
+			<div className="mx-auto w-full max-w-[1320px] px-6">
+				<Nav />
+			</div>
+			<div className="mx-auto flex w-full max-w-[1320px] flex-1 px-6">
 				<Sidebar />
 				<main className="docs-content min-w-0 flex-1 px-6 py-8 sm:px-10 lg:px-16">
 					<div className="mx-auto max-w-3xl">{children}</div>
